Python/Recommendation_System/requirements.txt
flask
requests
nltk
sentence-transformers
pymongo
scikit-learn
faiss-cpu